Added the rest of the major components inside of where the App component is rendering them to test if it works, and it successfully rendered. Side note: in the App components return statement with the rest of the components, I have commented out the Search component for now, because it's styling needs work and it's a low priority at the moment.
Also edited the Search component to have comments indicating the styled components that need to be fixed as a reference for the future.

Finally, after researching and testing the difference between ReactDOM.render and ReactDOM.createRoot, I was able to successfully switch over to ReactDOM.createRoot and root.render, which I think we should keep.